# frontend/docs/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Orient-advertise-backend API",
    "description" : "Multi-module Spring Boot platform for TV-box content delivery — device management, content uploads + transcoding, playlists, scheduling, remote actions, incident lifecycle, proof-of-play, aggregated reporting, Excel export, and external partner integrations.\n\n**Authentication.** Most endpoints require a JWT access token (`Authorization: Bearer <token>`) obtained from `POST /auth/login`. External integration endpoints under `/service/external/**` use an `X-API-Key` header instead — see the *External* group.\n\n**Sensitive endpoints.** Operations marked with `[SENSITIVE]` in their summary affect production state, expose user data, or irreversibly delete records. Audit logging is enabled and the operator's username is recorded.\n",
    "contact" : {
      "name" : "orientadvertise.services"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "version" : "1.0.55"
  },
  "servers" : [ {
    "url" : "/",
    "description" : "Current host"
  }, {
    "url" : "https://api.example.com",
    "description" : "Production"
  } ],
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
...
  "paths" : {
...
    "/api/devices/{id}/remote" : {
      "get" : {
        "tags" : [ "Remote Control" ],
        "summary" : "Current remote session for a device",
        "operationId" : "current",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The live session (no viewerTicket)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RemoteSessionResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "No live session for this device",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RemoteSessionResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "Caller is VIEWER/ADVERTISER",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RemoteSessionResponse"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Remote Control", "Sensitive" ],
        "summary" : "[SENSITIVE] Start a remote view/control session",
        "operationId" : "start",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/StartRemoteSessionRequest"
              }
            }
          }
        },
        "responses" : {
          "404" : {
            "description" : "Device not found, soft-deleted, or out of scope",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RemoteSessionResponse"
                }
              }
            }
          },
          "503" : {
            "description" : "Remote control is disabled on this server",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RemoteSessionResponse"
                }
              }
            }
          },
          "201" : {
            "description" : "Session created",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RemoteSessionResponse"
                }
              }
            }
          },
          "422" : {
            "description" : "Device reported capability.supported == false",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RemoteSessionResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "Caller is VIEWER/ADVERTISER",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RemoteSessionResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "A PENDING/ACTIVE session already exists",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RemoteSessionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/devices/{id}/remote/{sessionKey}/ack" : {
      "post" : {
        "tags" : [ "Remote Control" ],
        "summary" : "Device acknowledges a remote session (READY / FAILED / ENDED)",
        "operationId" : "ack",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "sessionKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RemoteSessionAckRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Ack recorded",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AckResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "Token belongs to a different device",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AckResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Unknown session, or it belongs to another device",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AckResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "Session is already terminal",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AckResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/devices/{id}/heartbeat" : {
      "post" : {
        "tags" : [ "device-controller" ],
        "operationId" : "heartbeat",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/HeartbeatRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/HeartbeatResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/assignments/{id}/confirm" : {
      "post" : {
        "tags" : [ "assignment-controller" ],
        "summary" : "Confirm a draft assignment with optional device exclusions or inclusions",
        "description" : "Send `excludedDeviceIds` to confirm against the full target scope minus those devices. Send `includedDeviceIds` to confirm against ONLY those devices (the backend derives the complement set as exclusions — useful when a target has more than the 200-device preview cap and the FE can't enumerate the full set client-side). The two fields are mutually exclusive: sending both returns 400.",
        "operationId" : "confirm",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ConfirmRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "409" : {
            "description" : "Time overlap with an existing CONFIRMED assignment (retry with replaceConflicting:true to supersede it — see details.conflicts), the assignment is not in DRAFT status, or the playlist has no items",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AssignmentResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Unknown assignment or excluded device",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AssignmentResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "Both excludedDeviceIds and includedDeviceIds set, or includedDeviceIds contains ids outside the assignment target scope",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AssignmentResponse"
                }
              }
            }
          },
          "200" : {
            "description" : "Assignment confirmed",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AssignmentResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/devices/{id}/remote/{sessionKey}" : {
      "delete" : {
        "tags" : [ "Remote Control", "Sensitive" ],
        "summary" : "[SENSITIVE] Stop a remote view/control session",
        "operationId" : "stop",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "sessionKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "404" : {
            "description" : "Unknown session, or it belongs to another device"
          },
          "204" : {
            "description" : "Stopped (or already terminal)"
          },
          "403" : {
            "description" : "Caller is VIEWER/ADVERTISER"
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "StartRemoteSessionRequest" : {
        "type" : "object",
        "properties" : {
          "viewOnly" : {
            "type" : "boolean"
          }
        }
      },
      "CapabilityDto" : {
        "type" : "object",
        "properties" : {
          "supported" : {
            "type" : "boolean"
          },
          "input" : {
            "type" : "string"
          },
          "transport" : {
            "type" : "string"
          },
          "maxWidth" : {
            "type" : "integer",
            "format" : "int32"
          },
          "maxHeight" : {
            "type" : "integer",
            "format" : "int32"
          },
          "reportedAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "RemoteSessionResponse" : {
        "type" : "object",
        "properties" : {
          "sessionId" : {
            "type" : "string"
          },
          "deviceId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "status" : {
            "type" : "string"
          },
          "relayUrl" : {
            "type" : "string"
          },
          "viewerTicket" : {
            "type" : "string"
          },
          "expiresAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "viewOnly" : {
            "type" : "boolean"
          },
          "deliveredVia" : {
            "type" : "string"
          },
          "capability" : {
            "$ref" : "#/components/schemas/CapabilityDto"
          }
        }
      },
      "RemoteSessionAckRequest" : {
        "required" : [ "status" ],
        "type" : "object",
        "properties" : {
          "status" : {
            "type" : "string",
            "enum" : [ "READY", "FAILED", "ENDED" ]
          },
          "width" : {
            "type" : "integer",
            "format" : "int32"
          },
          "height" : {
            "type" : "integer",
            "format" : "int32"
          },
          "error" : {
            "type" : "string"
          },
          "reason" : {
            "type" : "string"
          }
        }
      },
      "AckResponse" : {
        "type" : "object",
        "properties" : {
          "sessionId" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          }
        }
      },
...
      "HeartbeatRequest" : {
        "type" : "object",
        "properties" : {
          "contentVersion" : {
            "type" : "string"
          },
          "volume" : {
            "type" : "integer",
            "format" : "int32"
          },
          "remote" : {
            "$ref" : "#/components/schemas/RemoteCapabilityDto"
          }
        }
      },
      "RemoteCapabilityDto" : {
        "type" : "object",
        "properties" : {
          "supported" : {
            "type" : "boolean"
          },
          "input" : {
            "type" : "string"
          },
          "transport" : {
            "type" : "string"
          },
          "maxWidth" : {
            "type" : "integer",
            "format" : "int32"
          },
          "maxHeight" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "DesiredRemoteSessionDto" : {
        "type" : "object",
        "properties" : {
          "sessionId" : {
            "type" : "string"
          },
          "relayUrl" : {
            "type" : "string"
          },
          "agentTicket" : {
            "type" : "string"
          },
          "expiresAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "viewOnly" : {
            "type" : "boolean"
          },
          "maxWidth" : {
            "type" : "integer",
            "format" : "int32"
          },
          "maxFps" : {
            "type" : "integer",
            "format" : "int32"
          },
          "bitRate" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "HeartbeatResponse" : {
        "type" : "object",
        "properties" : {
          "deviceId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "status" : {
            "type" : "string"
          },
          "serverTime" : {
            "type" : "string",
            "format" : "date-time"
          },
          "pendingActions" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PendingActionDto"
            }
          },
          "expectedContentVersion" : {
            "type" : "string"
          },
          "syncRequired" : {
            "type" : "boolean"
          },
          "desiredVolume" : {
            "type" : "integer",
            "format" : "int32"
          },
          "syncGroupId" : {
            "type" : "string"
          },
          "desiredRemoteSession" : {
            "$ref" : "#/components/schemas/DesiredRemoteSessionDto"
          }
        }
      },
      "PendingActionDto" : {
        "type" : "object",
        "properties" : {
          "actionId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "actionType" : {
            "type" : "string"
          },
          "payload" : {
            "type" : "string"
          },
          "issuedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "expiresAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "AssignmentResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "playlistId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "targetType" : {
            "type" : "string"
          },
          "targetId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "priority" : {
            "type" : "integer",
            "format" : "int32"
          },
          "startTime" : {
            "type" : "string",
            "format" : "date-time"
          },
          "endTime" : {
            "type" : "string",
            "format" : "date-time"
          },
          "status" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "ConfirmRequest" : {
        "type" : "object",
        "properties" : {
          "excludedDeviceIds" : {
            "type" : "array",
            "items" : {
              "type" : "integer",
              "format" : "int64"
            }
          },
          "includedDeviceIds" : {
            "type" : "array",
            "items" : {
              "type" : "integer",
              "format" : "int64"
            }
          },
          "reason" : {
            "type" : "string"
          },
          "replaceConflicting" : {
            "type" : "boolean"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "description" : "JWT access token from `POST /auth/login`. Roles encoded in the token: ADMIN, OPERATOR, VIEWER, ADVERTISER.",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      },
      "apiKeyAuth" : {
        "type" : "apiKey",
        "description" : "External-integration API key. Issued by ADMIN via `POST /service/admin/service-keys`. Rate-limited 100 req/hr per key. Revocation is immediate — no in-memory cache widens the window.",
        "name" : "X-API-Key",
        "in" : "header"
      }
    }
  }
}